requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
mysqlclient==2.2.4
gunicorn==21.2.0
bcrypt==4.1.2
requests==2.31.0
pytz==2024.1
